+++ Models/Dto/ContactoPaciente/UpdateContactoPacienteDto.cs
@@ -9,8 +9,8 @@ namespace Satizen_Api.Models.Dto.ContactoPaciente
     {
         public int idContacto { get; set; }
         public int idPaciente { get; set; }
-        public int celularPaciente { get; set; }
-        public int celularAcompananteP { get; set; }
+        public long celularPaciente { get; set; }
+        public long celularAcompananteP { get; set; }
         public string estadoContacto { get; set; }
     }
 }
